+++ azext_iot/_help.py
@@ -200,9 +200,15 @@ helps[
       text: >
         az iot hub device-identity update -d {device_id} -n {iothub_name}
         --set capabilities.iotEdge=true
+    - name: Turn on edge capabilities for device using convenience argument.
+      text: >
+        az iot hub device-identity update -d {device_id} -n {iothub_name} --ee
     - name: Disable device status
       text: >
         az iot hub device-identity update -d {device_id} -n {iothub_name} --set status=disabled
+    - name: Disable device status using convenience argument.
+      text: >
+        az iot hub device-identity update -d {device_id} -n {iothub_name} --status disabled
     - name: In one command
       text: >
         az iot hub device-identity update -d {device_id} -n {iothub_name}