+++ azext_iot/tests/test_iot_ext_int.py
@@ -444,8 +444,55 @@ class TestIoTHubDevices(IoTLiveScenarioTest):
         )
 
         self.cmd(
-            '''iot hub device-identity update -d {} -n {} -g {} --set authentication.symmetricKey.primaryKey=""
-                    authentication.symmetricKey.secondaryKey=""'''.format(
+            "iot hub device-identity update -d {} -n {} -g {} --ee {} --auth-method {}"
+            .format(device_ids[0], LIVE_HUB, LIVE_RG, False, 'x509_ca'),
+            checks=[
+                self.check("deviceId", device_ids[0]),
+                self.check("status", "enabled"),
+                self.check("capabilities.iotEdge", False),
+                self.check("authentication.symmetricKey.primaryKey", None),
+                self.check("authentication.symmetricKey.secondaryKey", None),
+                self.check("authentication.x509Thumbprint.primaryThumbprint", None),
+                self.check("authentication.x509Thumbprint.secondaryThumbprint", None),
+                self.check("authentication.type", 'certificateAuthority')
+            ]
+        )
+
+        self.cmd(
+            "iot hub device-identity update -d {} -n {} -g {} --status-reason {}"
+            .format(device_ids[0], LIVE_HUB, LIVE_RG, 'TestStatusReason'),
+            checks=[
+                self.check("deviceId", device_ids[0]),
+                self.check("statusReason", 'TestStatusReason'),
+            ]
+        )
+
+        self.cmd(
+            "iot hub device-identity update -d {} -n {} -g {} --ee {} --status {}"
+            " --status-reason {} --auth-method {} --ptp {} --stp {}"
+            .format(device_ids[0], LIVE_HUB, LIVE_RG, False, 'enabled',
+                    'StatusReasonUpdated', 'x509_thumbprint', PRIMARY_THUMBPRINT, SECONDARY_THUMBPRINT),
+            checks=[
+                self.check("deviceId", device_ids[0]),
+                self.check("status", "enabled"),
+                self.check("capabilities.iotEdge", False),
+                self.check("statusReason", 'StatusReasonUpdated'),
+                self.check("authentication.x509Thumbprint.primaryThumbprint", PRIMARY_THUMBPRINT),
+                self.check("authentication.x509Thumbprint.secondaryThumbprint", SECONDARY_THUMBPRINT),
+            ]
+        )
+
+        self.cmd("iot hub device-identity update -d {} -n {} -g {} --auth-method {}"
+                 .format(device_ids[0], LIVE_HUB, LIVE_RG, 'x509_thumbprint'),
+                 expect_failure=True)
+
+        self.cmd("iot hub device-identity update -d {} -n {} -g {} --auth-method {} --pk {}"
+                 .format(device_ids[0], LIVE_HUB, LIVE_RG, 'shared_private_key', '123'),
+                 expect_failure=True)
+
+        self.cmd(
+            '''iot hub device-identity update -d {} -n {} -g {} --primary-key=""
+                    --secondary-key=""'''.format(
                 edge_device_ids[1], LIVE_HUB, LIVE_RG
             ),
             checks=[